backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
authlib==1.3.1
cryptography==43.0.1
python-jose==3.3.0
argon2-cffi==23.1.0
passlib==1.7.4
celery==5.4.0
redis==5.0.1
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.2
itsdangerous==2.2.0
pytest==8.3.3
pytest-asyncio==0.23.8
pandas==2.2.2
openpyxl==3.1.5
